# packages/contracts/openapi/api.openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FieldError"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "requestId": {
            "title": "Requestid",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "requestId"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "FieldError": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "field": {
            "title": "Field",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "field",
          "code",
          "message"
        ],
        "title": "FieldError",
        "type": "object"
      },
...
      "RunState": {
        "enum": [
          "initializing",
          "running",
          "stopping",
          "finished",
          "failed",
          "aborted"
        ],
        "title": "RunState",
        "type": "string"
      },
      "RunStopResponse": {
        "properties": {
          "duplicate": {
            "title": "Duplicate",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/RunState"
          },
          "stopRequestedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stoprequestedat"
          }
        },
        "required": [
          "id",
          "state",
          "duplicate"
        ],
        "title": "RunStopResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "SurgePilot API",
    "version": "0.1.0",
    "x-surgepilot-error-codes": [
      "ENV_GROUP_NAME_CONFLICT",
      "ENV_GROUP_IN_USE",
      "INVALID_FILENAME",
      "DEPENDENCY_FILE_NAME_CONFLICT",
      "PAYLOAD_TOO_LARGE",
      "STORAGE_UNAVAILABLE",
      "FILE_IN_USE",
      "LOAD_NODE_PUBLIC_ADMIN_REQUIRED",
      "LOAD_NODE_CONFLICT",
      "LOAD_NODE_BUSY",
      "LOAD_NODE_ACTION_NOT_ALLOWED",
      "LOAD_NODE_CREDENTIAL_REQUIRED",
      "LOAD_NODE_CREDENTIAL_INVALID",
      "LOAD_NODE_HOST_INVALID",
      "LOAD_NODE_RUNNER_HOME_INVALID",
      "LOAD_NODE_SSH_TIMEOUT",
      "LOAD_NODE_SSH_AUTH_FAILED",
      "LOAD_NODE_SSH_HOST_KEY_SCAN_FAILED",
      "LOAD_NODE_SSH_HOST_KEY_MISMATCH",
      "LOAD_NODE_SSH_HOST_KEY_UNTRUSTED",
      "LOAD_NODE_SSH_HOST_KEY_CHANGED",
      "LOAD_NODE_SSH_UNREACHABLE",
      "LOAD_NODE_RUNNER_HOME_UNWRITABLE",
      "LOAD_NODE_PYTHON_MISSING",
      "LOAD_NODE_JAVA_MISSING",
      "LOAD_NODE_TAURUS_MISSING",
      "LOAD_NODE_JMETER_MISSING",
      "LOAD_NODE_INIT_FAILED",
      "CREDENTIAL_DECRYPT_FAILED"
    ]
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/runs/{runId}/stop": {
      "post": {
        "operationId": "stopRun",
        "parameters": [
          {
            "in": "path",
            "name": "runId",
            "required": true,
            "schema": {
              "title": "Runid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-workspace-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Workspace-Id"
            }
          },
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": true,
            "schema": {
              "title": "X-Csrf-Token",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "surgepilot_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "SurgePilot Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunStopResponse"
                }
              }
            },
            "description": "OK"
          },
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunStopResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Stop Run",
        "tags": [
          "runs"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "/api"
    }
  ]
}